Refactor code to avoid function duplication
Previously, the crossOut() function was defined twice in the code, which made it harder to maintain. This commit refactors the code to define the crossOut() function only once and pass a reference to it as the second argument to the on() method. This makes the code more concise and easier to read. No changes were made to the functionality of the code.